% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_trend.R
\encoding{UTF-8}
\name{create_vas_plots_comparison}
\alias{create_vas_plots_comparison}
\title{Create Country-wise VAS Plots}
\usage{
create_vas_plots_comparison(data, lang = "ja")
}
\arguments{
\item{data}{Long-format data}

\item{lang}{Language code}
}
\value{
ggplot object
}
\description{
Creates Visual Analogue Scale (VAS) plots comparing data between countries
}
\concept{visualization}
