% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_environment}
\alias{initialize_environment}
\title{環境設定の初期化}
\usage{
initialize_environment(
  verbose = FALSE,
  install_packages = TRUE,
  force_cairo = FALSE
)
}
\arguments{
\item{verbose}{デバッグ情報の表示フラグ}

\item{install_packages}{不足パッケージの自動インストールフラグ}

\item{force_cairo}{Cairoデバイスの強制使用フラグ}
}
\value{
list 設定結果を含むリスト
\item{success}{bool 全体の成功状態}
\item{locale}{character 設定されたロケール}
\item{graphics_device}{character 設定されたグラフィックデバイス}
\item{messages}{character vector 処理中のメッセージ}
}
\description{
システムの環境設定、ロケール、グラフィックデバイスの初期化を行う
}
