% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_trend.R
\name{perform_country_ttest}
\alias{perform_country_ttest}
\title{国別のt検定実施}
\usage{
perform_country_ttest(data, lang = "ja")
}
\arguments{
\item{data}{長形式データ}

\item{lang}{言語コード}
}
\value{
t検定結果のデータフレーム
}
\description{
国別のt検定実施
}
