% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_common_theme}
\alias{apply_common_theme}
\title{プロットの共通テーマを設定する関数}
\usage{
apply_common_theme(base_size = 14, rotate_x_labels = FALSE)
}
\arguments{
\item{base_size}{ベースとなるフォントサイズ（デフォルト：14）}

\item{rotate_x_labels}{X軸ラベルを45度回転するかどうか（デフォルト：FALSE）}

\item{p}{ggplotオブジェクト}
}
\value{
テーマが適用されたggplotオブジェクト
}
\description{
ggplotオブジェクトに対して、統一されたテーマ設定を適用します
}
