% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_trend.R
\name{create_vas_plot}
\alias{create_vas_plot}
\title{最適化されたVASプロットの作成}
\usage{
create_vas_plot(data, lang = "ja", country_filter = NULL)
}
\arguments{
\item{data}{長形式データ}

\item{lang}{言語コード}

\item{country_filter}{特定の国のデータのみを表示する場合の国名}
}
\value{
ggplotオブジェクト
}
\description{
最適化されたVASプロットの作成
}
