% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_trend.R
\name{create_vas_plots_comparison}
\alias{create_vas_plots_comparison}
\title{国別VASプロットの作成}
\usage{
create_vas_plots_comparison(data, lang = "ja")
}
\arguments{
\item{data}{長形式データ}

\item{lang}{言語コード}
}
\value{
ggplotオブジェクト
}
\description{
国別VASプロットの作成
}
