% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_config.R
\name{get_error_message}
\alias{get_error_message}
\title{エラーメッセージの取得}
\usage{
get_error_message(error_key, lang = getOption("movie_analysis.lang", "ja"))
}
\arguments{
\item{error_key}{エラーキー}

\item{lang}{言語コード}
}
\value{
character エラーメッセージ
}
\description{
エラーメッセージの取得
}
