% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_trend.R
\name{main_basic_trend}
\alias{main_basic_trend}
\title{メイン実行関数の改善版}
\usage{
main_basic_trend(
  input_file,
  output_dir,
  lang = "ja",
  verbose = FALSE,
  benchmark = FALSE
)
}
\arguments{
\item{input_file}{入力ファイルパス}

\item{output_dir}{出力ディレクトリ}

\item{lang}{言語コード}

\item{verbose}{デバッグ情報の表示フラグ}

\item{benchmark}{ベンチマーク実行フラグ}
}
\description{
メイン実行関数の改善版
}
